refatoracao para module
